+++ sandbox/apps/sandbox/index.js
@@ -91,5 +91,7 @@ module.exports = {
       ],
       next: '/confirm'
     },
+    '/session-timeout': {},
+    '/exit': {}
   }
 };